+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 ### 2.0.1: 2025-06-15
 
 * Simplify memory updated status text
+* Prefer English
+* Add foreign language support
 
 ### 2.0.0: 2025-06-15